+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/injection/RawArgs.java
@@ -34,7 +34,6 @@ import java.lang.annotation.Target;
  * <p>
  * This should only be used on {@code String[]}
  *
- *
  * @since 1.2.0
  */
 @Target(ElementType.METHOD)