+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - Argument builders now use generics to preserve their type when using functions from the base class.
 
+## [Unreleased]
+
+### Changed
+ - Moved the parser injector registry into CommandManager and added injection to CommandContext
+
 ## [1.2.0] - 2020-12-07
 
 ### Added